+++ packages/utils/index.ts
@@ -14,4 +14,4 @@ export * from './types'
 export * from './typescript'
 export * from './throttleByRaf'
 export * from './easings'
-export * from './number'
+export * from './numbers'
